executor: migrate test-infra to testify for executor_test.testSerialSuite2
